+++ lib/src/omemo/ratchet_data.dart
@@ -5,8 +5,6 @@ class OmemoRatchetData {
     this.jid,
     this.id,
     this.ratchet,
-    this.added,
-    this.replaced,
   );
 
   /// The JID we have the ratchet with.
@@ -17,10 +15,4 @@ class OmemoRatchetData {
 
   /// The actual double ratchet to commit.
   final OmemoDoubleRatchet ratchet;
-
-  /// Indicates whether the ratchet has just been created (true) or just modified (false).
-  final bool added;
-
-  /// Indicates whether the ratchet has been replaced (true) or not.
-  final bool replaced;
 }